New PNM in progress
* Refactor code in 'src/backend/executor/execGpmon.c' to enable new changes and report start & finish states
* Add 'pernode' to a list of tables that should be modified by harvester process (so that the old tuples go to '_history' table)

+++ src/include/gpmon/gpmon.h
@@ -238,8 +238,8 @@ typedef struct gpmon_planmetric_key_t {
 typedef struct gpmon_planmetric_t {
 	gpmon_planmetric_key_t key;
 	int32 node;			/* GPDB node code */
-	uint64 t_start;		/* start processing of the node */
-	uint64 t_finish;	/* end processing of the node */
+	int32 t_start;		/* start processing of the node */
+	int32 t_finish;	/* end processing of the node */
 } gpmon_planmetric_t;
